Extract out misk client interfaces and classes into the api module for cross-framework support
