+++ misk-api/api/misk-api.api
@@ -5,6 +5,45 @@ public abstract interface class misk/api/HttpRequest {
 	public abstract fun setRequestHeaders (Lokhttp3/Headers;)V
 }
 
+public final class misk/client/ClientAction {
+	public fun <init> (Ljava/lang/String;Lkotlin/reflect/KFunction;)V
+	public fun <init> (Ljava/lang/String;Lkotlin/reflect/KFunction;Ljava/util/List;Lkotlin/reflect/KType;)V
+	public final fun component1 ()Ljava/lang/String;
+	public final fun component2 ()Lkotlin/reflect/KFunction;
+	public final fun component3 ()Ljava/util/List;
+	public final fun component4 ()Lkotlin/reflect/KType;
+	public final fun copy (Ljava/lang/String;Lkotlin/reflect/KFunction;Ljava/util/List;Lkotlin/reflect/KType;)Lmisk/client/ClientAction;
+	public static synthetic fun copy$default (Lmisk/client/ClientAction;Ljava/lang/String;Lkotlin/reflect/KFunction;Ljava/util/List;Lkotlin/reflect/KType;ILjava/lang/Object;)Lmisk/client/ClientAction;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getFunction ()Lkotlin/reflect/KFunction;
+	public final fun getName ()Ljava/lang/String;
+	public final fun getParameterTypes ()Ljava/util/List;
+	public final fun getReturnType ()Lkotlin/reflect/KType;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
+public abstract interface class misk/client/ClientNetworkChain {
+	public abstract fun getAction ()Lmisk/client/ClientAction;
+	public abstract fun getRequest ()Lokhttp3/Request;
+	public abstract fun proceed (Lokhttp3/Request;)Lokhttp3/Response;
+}
+
+public abstract interface class misk/client/ClientNetworkInterceptor {
+	public abstract fun intercept (Lmisk/client/ClientNetworkChain;)Lokhttp3/Response;
+}
+
+public abstract interface class misk/client/ClientNetworkInterceptor$Factory {
+	public abstract fun create (Lmisk/client/ClientAction;)Lmisk/client/ClientNetworkInterceptor;
+}
+
+public final class misk/client/NetworkInterceptorWrapper : okhttp3/Interceptor {
+	public fun <init> (Lmisk/client/ClientAction;Lmisk/client/ClientNetworkInterceptor;)V
+	public final fun getAction ()Lmisk/client/ClientAction;
+	public final fun getInterceptor ()Lmisk/client/ClientNetworkInterceptor;
+	public fun intercept (Lokhttp3/Interceptor$Chain;)Lokhttp3/Response;
+}
+
 public abstract interface class misk/scope/ActionScoped {
 	public static final field Companion Lmisk/scope/ActionScoped$Companion;
 	public abstract fun get ()Ljava/lang/Object;
