+++ misk/api/misk.api
@@ -418,23 +418,6 @@ public abstract interface class misk/client/CallFactoryWrapper {
 	public abstract fun wrap (Lmisk/client/ClientAction;Lokhttp3/Call$Factory;)Lokhttp3/Call$Factory;
 }
 
-public final class misk/client/ClientAction {
-	public fun <init> (Ljava/lang/String;Lkotlin/reflect/KFunction;Ljava/util/List;Lkotlin/reflect/KType;)V
-	public final fun component1 ()Ljava/lang/String;
-	public final fun component2 ()Lkotlin/reflect/KFunction;
-	public final fun component3 ()Ljava/util/List;
-	public final fun component4 ()Lkotlin/reflect/KType;
-	public final fun copy (Ljava/lang/String;Lkotlin/reflect/KFunction;Ljava/util/List;Lkotlin/reflect/KType;)Lmisk/client/ClientAction;
-	public static synthetic fun copy$default (Lmisk/client/ClientAction;Ljava/lang/String;Lkotlin/reflect/KFunction;Ljava/util/List;Lkotlin/reflect/KType;ILjava/lang/Object;)Lmisk/client/ClientAction;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getFunction ()Lkotlin/reflect/KFunction;
-	public final fun getName ()Ljava/lang/String;
-	public final fun getParameterTypes ()Ljava/util/List;
-	public final fun getReturnType ()Lkotlin/reflect/KType;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
 public abstract interface class misk/client/ClientApplicationInterceptorFactory {
 	public abstract fun create (Lmisk/client/ClientAction;)Lokhttp3/Interceptor;
 }
@@ -455,20 +438,6 @@ public final class misk/client/ClientMetricsInterceptor : okhttp3/Interceptor {
 
 public final class misk/client/ClientMetricsInterceptor$Factory {
 	public final fun create (Ljava/lang/String;)Lmisk/client/ClientMetricsInterceptor;
-}
-
-public abstract interface class misk/client/ClientNetworkChain {
-	public abstract fun getAction ()Lmisk/client/ClientAction;
-	public abstract fun getRequest ()Lokhttp3/Request;
-	public abstract fun proceed (Lokhttp3/Request;)Lokhttp3/Response;
-}
-
-public abstract interface class misk/client/ClientNetworkInterceptor {
-	public abstract fun intercept (Lmisk/client/ClientNetworkChain;)Lokhttp3/Response;
-}
-
-public abstract interface class misk/client/ClientNetworkInterceptor$Factory {
-	public abstract fun create (Lmisk/client/ClientAction;)Lmisk/client/ClientNetworkInterceptor;
 }
 
 public final class misk/client/ClientNetworkInterceptorsModule : misk/inject/KAbstractModule {
