+++ misk/src/main/kotlin/misk/client/ClientInvocationHandler.kt
@@ -126,15 +126,6 @@ internal class ClientInvocationHandler(
   }
 }
 
-internal class NetworkInterceptorWrapper(
-  val action: ClientAction,
-  val interceptor: ClientNetworkInterceptor
-) : Interceptor {
-  override fun intercept(chain: Interceptor.Chain): okhttp3.Response {
-    return interceptor.intercept(RealClientNetworkChain(chain, action))
-  }
-}
-
 private val KFunction<*>.isRetrofitMethod: Boolean
   get() {
     return annotations.any {
